fixed all the warnings

+++ Moments/Libraries/AFNetworking/FBEncryptorAES.m
@@ -153,7 +153,8 @@
 {
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
-        srand(time(NULL));
+        
+        srand((unsigned int)time(NULL));
     });
 
     char cIv[FBENCRYPT_BLOCK_SIZE];
